+++ bundle/src/main/java/com/adobe/acs/commons/util/impl/ReflectionUtil.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.util;
+package com.adobe.acs.commons.util.impl;
 
 import com.adobe.acs.commons.util.impl.ValueMapTypeConverter;
 import com.day.cq.commons.inherit.InheritanceValueMap;
